test(objstore): bound bench seed to uint64
`fmt.Sprintf("%040x", 0xaaaaaaaa)` in the symbolic-HEAD
fixture for the bench overflows the untyped-int upper bound
on 32-bit platforms (max int32 is 0x7FFF_FFFF; the seed is
0xAAAA_AAAA). Surfaced by the new `test-386` cell.

Cast through `uint64` so the literal fits regardless of host
word size. The rendered hex is unchanged.

+++ internal/objstore/store_bench_test.go
@@ -317,7 +317,7 @@ func makeLooseRefsFixture(b *testing.B, n int) string {
 	// the numbered branches each get a distinct synthetic OID derived
 	// from their index. Distinct OIDs are not strictly required by the
 	// iterator but keep the fixture honest.
-	mainOID := fmt.Sprintf("%040x", 0xaaaaaaaa)
+	mainOID := fmt.Sprintf("%040x", uint64(0xaaaaaaaa))
 	if err := os.WriteFile(filepath.Join(gitDir, "refs", "heads", "main"),
 		[]byte(mainOID+"\n"), 0o644); err != nil {
 		b.Fatal(err)
